Fix gem not working after httpi and savon updates
In the recent update of savon gem from 2.12 to 2.14, the dependency on httpi is loosened from  ~> 2.3 to >= 2.4.5. It means that httpi can get updated to version 3+.

This change crashes gus_bir1. The proposed change fixes it.

It is not however researched thoroughly, there might other areas of the gem that are still affected.